backend/requirements.txt
# Django & DRF Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Database & ORM
psycopg2-binary==2.9.9
django-extensions==3.2.3

# Authentication & Security
djangorestframework-simplejwt==5.3.0
cryptography==41.0.8
pycryptodome==3.19.0

# Web3 & Blockchain
web3==6.12.0
eth-account==0.9.0

# Financial & Payment Processing
stripe==7.7.0
plaid-python==11.0.0

# Utilities
redis==5.0.1
celery==5.3.4
python-decouple==3.8

# Development
django-debug-toolbar==4.2.0
pytest-django==4.6.0
factory-boy==3.3.0